reorganize to conventional holos layout
This patch re-organizes the layout to match the holos conventional
layout.  We don't need to modify any files, we only relocate them.

./my-chart moves to ./components/my-chart/vendor/0.1.0/my-chart for use
with the holos `#Helm` component kind.  Holos supports vendored charts
committed to the repository.

./my-values moves to ./components/my-chart/my-values so the holos
component can pass the correct values files to `helm template`.

./appsets moves to ./not-used because we don't need any of them.  We'll
migrate the entire behavior of the
`not-used/appsets/4-final/all-my-envs-appset-with-version.yaml`
ApplicationSet resource to holos components and fully hydrated
Application resources.

./example-apps is used in an unrelated portion of the original blog post
so we move it to ./not-used