reorganize to conventional holos layout
This patch moves files without modification to match the holos
conventional layout.

./my-chart moves to ./components/my-chart/vendor/0.1.0/my-chart for use
with the holos `#Helm` component kind.  Holos supports vendored charts
committed to the repository.

./my-values moves to ./components/my-chart/my-values so the holos
component can pass the correct values files to `helm template`.  Value
files will be passed to Helm in the same order as previously.

./appsets moves to ./not-used because Holos generates an Application for
each component using ComponentConfig.

./example-apps is used in an unrelated portion of the original blog post
so we move it to ./not-used