cue: add schema constraints on each config.json file
One of the problems with the original blog post is any old data can be
thrown into the config.json files and the problem won't be caught until
ArgoCD tries to deploy the helm chart, far too late in the process.
Engineer cycle time is very slow as a result.

Holos offers the ability the check constraints quickly.

Result:

For example, if we try and change `staging/asia/config.json` to type "staging"
we immediately get an error indicating the only two valid values are "prod" and
"non-prod"

    CUE_EXPERIMENT=embed holos cue export ./config/environments --out=yaml

    could not run: config."staging/asia/config.json".type: 2 errors in empty disjunction: (and 2 more errors)
    config."staging/asia/config.json".type: 2 errors in empty disjunction:
    config."staging/asia/config.json".type: conflicting values "non-prod" and "staging":
        /Users/jeff/Holos/multi-sources-example/config/environments/environments.cue:12:28
        /Users/jeff/Holos/multi-sources-example/config/environments/environments.cue:18:18
        staging/asia/config.json:4:13
    config."staging/asia/config.json".type: conflicting values "prod" and "staging":
        /Users/jeff/Holos/multi-sources-example/config/environments/environments.cue:12:28
        /Users/jeff/Holos/multi-sources-example/config/environments/environments.cue:18:9
        staging/asia/config.json:4:13