kustomize the namespace to the env
The rendered manifests do not have a metadata.namespace value set for
each rendered resource.

This patch fixes the problem by passing each helm template through
kustomize to patch the namespace into the resources.  This is similar to
how the ApplicationSet used prior to the migration used the Application
template.spec.destination.namespace field to set the namespace.