compose argocd application resources into every component
Without this patch the migration from the ApplicationSet isn't complete
because we have no ArgoCD Application resources. This patch uses Holos
and CUE to mix an ArgoCD Application resource into each Holos Component.

With this patch the migration from [v0.1.0] ApplicationSets and Helm
Hierarchies to Holos is complete.

The result is equivalent to the ApplicationSet we migrated from with
some added benefits.  Previously the ApplicationSet used a generator to
generate each application:

    spec:
      generators:
      - git:
          repoURL: https://github.com/kostis-codefresh/multi-sources-example.git
          revision: HEAD
          files:
          - path: "appsets/4-final/env-config/**/config.json"
      template:
        metadata:
          name: '{{.env}}'

This has been replaced by the Holos ComponentConfig feature which allows
us to mix anything into each Holos Component.  We use the feature to mix
in an Application resource to each component, resulting in a clearly
readable Application manifest.

For example:

    deploy/gitops/prod-us-my-chart-application.gen.yaml

    apiVersion: argoproj.io/v1alpha1
    kind: Application
    metadata:
        labels:
            env: prod-us
        name: prod-us-my-chart
        namespace: argocd
    spec:
        destination:
            server: https://kubernetes.default.svc
        project: default
        source:
            path: deploy/environments/prod-us/components/my-chart
            repoURL: https://github.com/holos-run/multi-sources-example.git
            targetRevision: main

This patch also manages the argocd Namespace from the namespaces
component.

    holos render platform

    rendered namespaces for all environments in 319.823625ms
    rendered my-chart 0.2.0 for environment staging-eu in 339.621083ms
    rendered my-chart 0.1.0 for environment integration-gpu in 340.748833ms
    rendered my-chart 0.2.0 for environment integration-non-gpu in 342.011916ms
    rendered my-chart 0.1.0 for environment prod-eu in 342.984959ms
    rendered my-chart 0.2.0 for environment qa in 343.638917ms
    rendered my-chart 0.2.0 for environment staging-us in 344.777542ms
    rendered my-chart 0.1.0 for environment prod-us in 344.781458ms
    rendered platform in 344.838209ms

[v0.1.0]: https://github.com/holos-run/multi-sources-example/blob/v0.1.0/appsets/4-final/all-my-envs-appset-with-version.yaml