remove unused common-values.yaml
The helm value files hierarchy failed to unify previously.  This is
problem because unification is easier to understand, the primary goal of
the migration.

Holos makes it straight forward to grep the final values so we can
remove unused values blocking unification.

    (cd deploy/environments && grep -r replicas:)

    integration-gpu/components/my-chart/my-chart.gen.yaml:  replicas: 3
    prod-eu/components/my-chart/my-chart.gen.yaml:  replicas: 8
    integration-non-gpu/components/my-chart/my-chart.gen.yaml:  replicas: 3
    qa/components/my-chart/my-chart.gen.yaml:  replicas: 3
    staging-eu/components/my-chart/my-chart.gen.yaml:  replicas: 3
    prod-us/components/my-chart/my-chart.gen.yaml:  replicas: 10
    staging-us/components/my-chart/my-chart.gen.yaml:  replicas: 3

Even easier to see the only replicaCount values used are 10, 3, 8

    grep -r replicas: deploy/environments | awk '{print $NF}' | sort -u

    10
    3
    8

This patch starts an interative process of making the values files
unify.  First we'll identify and removed unused values which conflict
with values that override them in the hierarchy.  An unchanged deploy
directory verifies the data was not used.

Then we'll identify values that affect a small number of environments
and move the desired value to the most significant env-values.yaml file,
the one last in the hierarchy.

Once all values move to their most significant place we can remove them
from all places they're written over until the files unify.

Once unified, we can replace the use of valueFiles with Values to ensure
conflicts aren't introduced in the future.

Result:

No changes when the platform is rendered.

    holos render platform

    rendered namespaces for all environments in 319.274708ms
    rendered my-chart 0.1.0 for environment prod-us in 335.620958ms
    rendered my-chart 0.1.0 for environment integration-gpu in 336.445458ms
    rendered my-chart 0.1.0 for environment prod-eu in 337.783625ms
    rendered my-chart 0.2.0 for environment qa in 339.031625ms
    rendered my-chart 0.2.0 for environment staging-eu in 339.562166ms
    rendered my-chart 0.2.0 for environment integration-non-gpu in 340.072375ms
    rendered my-chart 0.2.0 for environment staging-us in 340.08325ms
    rendered platform in 340.529584ms